requirements.txt
streamlit
requests
beautifulsoup4
newspaper3k
huggingface_hub
openai
pandas

